iffcomp: correct chunk size to uint32, document WF vs EA-IFF 85 deviations
- backpatch.hp: size member, GetSize(), AddToSize(), SubtractFromSize() all
  use uint32_t; asserts updated from > 0 to != 0
- binary.cc: chunk size written as uint32_t (no cast needed now that
  GetSize() returns uint32_t); drop stale LP64 comment
- doc: reframe endianness as native-by-design (platform blob, not interchange
  file); add §5 WF vs EA-IFF 85 deviations table; fix chunk size type to
  uint32; add EA-IFF 85 spec reference

+++ wfsource/source/iffwrite/binary.cc
@@ -339,9 +339,7 @@ IffWriterBinary::exitChunk()
     // the 4 is the size of the IFF size field
 	_out->seekp( cs->GetPos(), std::ios::beg );
 
-    // IFF size field is always 4 bytes; on LP64 Linux `long` is 8 bytes,
-    // so use an explicit 32-bit integer here.
-    int32_t size = static_cast< int32_t >( cs->GetSize() );
+    uint32_t size = cs->GetSize();
     _out->write( reinterpret_cast< char const* >( &size ), 4 );
 	_out->seekp( pos, std::ios::beg );
